fixed: flickering when added and updated panel
This is related to design problem on GraphQL schema. So, please
reconsider the design. For example,

a. Do we need userId for Panel? (for update, remove whatever)
b. If not, how Panel, PanelList, DashBoard look like including
GraphQL schema and Apollo containers?

It's time to refactoring.

TODO: remove needs to be fixed.